wip: Add is_microop and microop_id to commit_if

+++ tests/regression/tmul/common.h
@@ -10,9 +10,9 @@
 typedef struct {
   uint32_t num_tasks;
   uint32_t size;
-  uint32_t A_addr;
-  uint32_t B_addr;
-  uint32_t C_addr;
+  uint64_t A_addr;
+  uint64_t B_addr;
+  uint64_t C_addr;
 } kernel_arg_t;
 
 #endif
